% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arrange_by_dim}
\alias{arrange_by_dim}
\title{Arrange \code{gp} well data relative to user supplied \code{start_corner}}
\usage{
arrange_by_dim(gp, dim)
}
\arguments{
\item{gp}{A \code{gp} object}

\item{dim}{Character. Either "row" or "col"}
}
\value{
A \code{gp} with \code{gp$well_data} arranged by dim and the non-dim (ie if dim
= "row", the 'non-dim' is "col") relative to the start corner (start corner
is at the top, farther from start corner is at bottom)
}
\description{
Arrange \code{gp} well data relative to user supplied \code{start_corner}
}
