% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplate.R
\name{as_gp}
\alias{as_gp}
\alias{as_gp.default}
\alias{as_gp.list}
\title{Coerce object to gp}
\usage{
as_gp(x, ...)

\method{as_gp}{default}(x, ...)

\method{as_gp}{list}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
Coerce object to gp
}
