% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate.R
\name{coordinate}
\alias{coordinate}
\title{Create a dimension pattern 'unit' of a section}
\usage{
coordinate(gp, type = c("row", "col"), margin)
}
\arguments{
\item{gp}{a \code{gp}}

\item{type}{Character. Either 'row' or 'col'.}

\item{margin}{Named list of integers defining the size of margins at each edge.}
}
\value{
a tibble
}
\description{
Create a dimension pattern 'unit' of a section
}
