% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_plate}
\alias{example_plate}
\title{A 8x12 matrix of random integers 1-10}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 12 columns.
}
\usage{
example_plate
}
\description{
A 8x12 matrix of random integers 1-10
}
\keyword{datasets}
