% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flip_dim}
\alias{flip_dim}
\title{Take a set of numbers and flip them around on a number line}
\usage{
flip_dim(gp, dim)
}
\arguments{
\item{gp}{A \code{gp}}

\item{dim}{Symbol. Column to flip. Should be the name of a column that exists in \code{gp$well_data}}
}
\value{
A vector.
}
\description{
Take a set of numbers and flip them around on a number line
}
\details{
This assumes the beginning is 1 and the end is maximum length of the
dimension. Note that this does NOT simply check the maximum value of the
supplied vector, but searches the supplied \code{gp} for a matching \code{ndim} column
that definitively tells how long the given dimension is supposed to be. This
is important, as sometimes a dimension is given that is longer than its
parent dimension, so the maximum numbers would not show in the vector.
}
