% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{gp}
\alias{gp}
\title{Make a gp object}
\usage{
gp(rows = NULL, cols = NULL, data = NULL, wells = NULL, tidy = FALSE)
}
\arguments{
\item{rows}{Numeric. The number of rows the plate should have.}

\item{cols}{Numeric. The number of columns the plate should have.}

\item{data}{An optional data.frame of well data the same dimensions as the
plate to be described}

\item{wells}{Numeric. The number of wells the plate has. If this is
specified, rows and cols must be null - they are inferred from common form
factors of plates.}

\item{tidy}{Are the data supplied tidy?}
}
\value{
a \code{gp} object
}
\description{
Make a gp object
}
\details{
A \code{gp} object has the following components:
\itemize{
\item \code{nrow}/\code{ncol}: Number of plate rows/cols. This is static and will not be
changed by adding layers.
\item \code{well_data}: Somewhat transient data used to define plotting coordinates
for layers. See below for more information.
\item \code{nrow_sec}/\code{ncols_sec}: The number of rows/cols of the current section.
When creating a plate, that number is the number of rows/cols of the plate
(the plate is the section). Can take on a 'mar' suffix, which specifies the
number including margins (if any)
\item \code{nrow_sec_par}/\code{ncol_sec_par}: The number of rows/cols of the parent
section. When creating a plate, it has no parent, so defaults to being its
own parent. Can take on a 'mar' suffix. See above.
}

\code{well_data} consists of many columns. The variable names can be broken down
as follows:
\itemize{
\item \code{row}/\code{col}: \code{row} is always the y axis, \code{col} is always the x axis. By
convention, plates start at 1, 1 in the top left corner.
\item \code{sec}: Short for 'section'. A section is a rectangular field of wells.
\code{sec} alone refers to the number of the section itself. \code{sec} combined with
\code{row} or \code{col} (eg \code{row_sec}) refers to the coordinates of a given well
relative to it's section corners, with the top left corner of a given section
always being (1, 1).
\item \code{par}: Short for 'parent'. These columns are all the data from the previous
layer.
\item \code{index}: These specify (usually) multiwell strips only defined in one
dimension. Together, \code{index_row} and \code{index_col} form
checkerboard-like patterns, where each intersection is a section.
This is a bit more complicated when \code{wrap = TRUE}, so the simile
does not hold for all cases.
}
}
\examples{

# If you specify wells, rows and columns are derived
# from a standard plate sizes:

gp(wells = 96)

# As such, you cannot use the wells argument
# if you want to create more exotic plates:

try(gp(wells = 102))

# For that, you'll need to specify wells and cols:

gp(rows = 6, cols = 17)

}
