% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_excise.R
\name{gp_excise}
\alias{gp_excise}
\title{Slice out a smaller gp from a gp via coordinates}
\usage{
gp_excise(gp, top, left, bottom, right)
}
\arguments{
\item{gp}{A \code{gp}}

\item{top, left, bottom, right}{Integer. The coordinates of the corners.
Remember that TOP left is (1, 1)}
}
\value{
A \code{gp} with top left coords = (1, 1)
}
\description{
Slice out a smaller gp from a gp via coordinates
}
\examples{
gplate::example_plate |> as_gp() |> gp_excise(2, 2, 4, 4)
}
