% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_section.R
\name{gp_sec}
\alias{gp_sec}
\title{Add a section to a \code{gp} object.}
\usage{
gp_sec(
  gp,
  id,
  rows = NULL,
  cols = NULL,
  flow = c("lrtb", "rltb", "lrbt", "rlbt"),
  padding = 0,
  margin = 0,
  break_sections = FALSE
)
}
\arguments{
\item{gp}{a \code{gp} object}

\item{id}{column of \code{gp$data} that defines the given section. Order is
determined by factor order, and will automatically undergo factor ordering
by \code{forcats::fct_inorder} if not a factor.}

\item{rows}{Numeric. Section height. If \code{NULL}, will fill width of parent
section.}

\item{cols}{Numeric. Section width. If \code{NULL}, will fill width of parent
section.}

\item{flow}{Determines how sections will be laid out. Default is \code{"lrtb"},
AKA left-right-top-bottom. Other valid values include}

\item{padding}{Border width INSIDE the section that will be unfilled. Can
take an argument of one (same border all around), two (top/bottom,
left/right), three (top, left/right, bottom), or four (top, right, bottom,
left).}

\item{margin}{Border width OUTSIDE the section that will remain unfilled. See
above for possible argument lengths.}

\item{break_sections}{If a section would need to be wrapped around the edge
of the parent section (default parent section is the plate), should it
instead shift the flow direction far enough so that it can fit without
wrapping? Will error if either \code{rows} or \code{cols} exceeds the parent
dimension.}
}
\value{

}
\description{
Add a section to a \code{gp} object.
}
