% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_serve.R
\name{gp_serve}
\alias{gp_serve}
\title{Extract useful, tidy data from a gp object}
\usage{
gp_serve(gp)
}
\arguments{
\item{gp}{a gp}
}
\value{
a \code{tibble}, with \code{.row} and \code{.col}, as well as any created section names and data values.
}
\description{
Extract useful, tidy data from a gp object
}
