% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_unserve.R
\name{gp_unserve}
\alias{gp_unserve}
\title{Turn a tidy \code{data.frame} into a \code{gp}}
\usage{
gp_unserve(x, row = ".row", col = ".col", nrow = NULL, ncol = NULL)
}
\arguments{
\item{x}{A tidy \code{data.frame} representation of plate data}

\item{row}{Character. The column that represents the rows. Should be numeric, with the lowest number representing the topmost row.}

\item{col}{Character. The name of the column that represents columns. Should be numeric, with the lowest number representing the leftmost column.}

\item{nrow}{Optional numeric, denoting the number of rows in the plate. If not supplied, will be imputed from the largest number in the 'row' column}

\item{ncol}{Optional numeric, denoting the number of columns in the plate. If not supplied, will be imputed from the largest number in the 'col' column}
}
\value{
a \code{gp}
}
\description{
An opposite to \code{serve()}
}
