% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_unwrap.R
\name{gp_unwrap}
\alias{gp_unwrap}
\title{Unwrap a plate into a single long lane}
\usage{
gp_unwrap(gp, flow)
}
\arguments{
\item{flow}{either "row" or "col"}

\item{wd}{Well data}

\item{nrow_sec}{The number of rows of a section, including margins}

\item{ncol_sec}{The number of columns of a section, including margins}
}
\value{
A tibble
}
\description{
Unwrap a plate into a single long lane
}
\details{
This returns a plate that IS nrow_sec tall or ncol_sec wide.
Therefore, it MUST be converted back into its normal plate form using its
cognate function
}
