% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_fwd}
\alias{is_fwd}
\title{Check if axis moves in the canonical direction}
\usage{
is_fwd(gp, dim)
}
\arguments{
\item{gp}{A \code{gp}}

\item{dim}{Character. A dimension, either "row" or "col".}
}
\value{
logical.
}
\description{
'Forwards' is thought of 'left to right' when thinking about moving across
columns and 'top to bottom' when moving across rows
}
