% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers<-}
\alias{layers<-}
\title{Layer setter generic}
\usage{
layers(x, position) <- value
}
\arguments{
\item{x}{Object to set layers of}

\item{position}{layer position to add or modify}

\item{value}{Value to set as layers}
}
\value{

}
\description{
Layer setter generic
}
