% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{layers}
\alias{layers}
\title{Layer accessor generic}
\usage{
layers(x, n = NULL)
}
\arguments{
\item{x}{Object to access layers of}

\item{n}{optional, layer position to access. If ommited, returns all layers}
}
\value{

}
\description{
Layer accessor generic
}
