% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplate.R
\name{new_gp}
\alias{new_gp}
\title{A constructor for gplates}
\usage{
new_gp(
  data = data.frame(),
  rows = integer(),
  cols = integer(),
  wells = integer(),
  options = list(col_dir = "top down", row_dir = "left right", excluded_border_width =
    0, row_or_col_first = "row")
)
}
\arguments{
\item{dimensions}{}
}
\value{

}
\description{
A constructor for gplates
}
