% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcr_plate}
\alias{pcr_plate}
\title{Data from a PCR experiment on a 384-well plate}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 16 rows and 24 columns.
}
\usage{
pcr_plate
}
\description{
Data from a PCR experiment on a 384-well plate
}
\keyword{datasets}
