% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate-class.R, R/layers.R, R/scrape.R,
%   R/serve.R, R/wells.R
\docType{class}
\name{Plate-class}
\alias{Plate-class}
\alias{layers,Plate-method}
\alias{layers<-,Plate-method}
\alias{spill,Plate-method}
\alias{scrape,Plate-method}
\alias{serve,Plate-method}
\alias{wells,Plate-method}
\alias{wells<-,Plate-method}
\title{An S4 class to represent microwell plate data}
\usage{
\S4method{layers}{Plate}(x, n = NULL)

\S4method{layers}{Plate}(x, position = NULL) <- value

\S4method{spill}{Plate}(x, n = NULL)

\S4method{scrape}{Plate}(x, positions = NULL, negate = FALSE)

\S4method{serve}{Plate}(x)

\S4method{wells}{Plate}(x)

\S4method{wells}{Plate}(x) <- value
}
\arguments{
\item{x}{a Plate object.}

\item{n}{optional, character name or numeric position of layer to access}

\item{position}{layer position to add or modify}

\item{value}{numeric. Number of wells to set.}

\item{positions}{numeric or character vector specifying the positions or names of layers to be removed.}

\item{negate}{logical. If TRUE, values will be kept instead of removed.}
}
\description{
An S4 class to represent microwell plate data
}
\section{Methods (by generic)}{
\itemize{
\item \code{layers}: accessor for plate layers

\item \code{layers<-}: add/overwrite a plate layer

\item \code{spill}: Alternative and cuter method to access layers

\item \code{scrape}: remove (or keep) layers by character or numeric vector.

\item \code{serve}: method to convert a plate object into a tidy tibble

\item \code{wells}: access number of wells on plate.

\item \code{wells<-}: set well number on a plate object.
}}

\section{Slots}{

\describe{
\item{\code{layers}}{a list of matrices}

\item{\code{meta}}{a named list of characteristics that refer to the whole plate}

\item{\code{wells}}{a numeric that describes how many wells are in the plate}

\item{\code{rows}}{a numeric denoting the number of rows on the plate}

\item{\code{cols}}{a numeric denoting the number of cols on the plate}
}}

