% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plate.R
\name{Plate}
\alias{Plate}
\title{Plate constructor}
\usage{
Plate(layers = list(), wells = NULL, meta = NULL)
}
\arguments{
\item{layers}{a matrix or dataframe, or a list of matrices/dataframes}

\item{wells}{optional numeric of number of wells the plate contains. Useful
when instantiating a plate without layers. Must match layer dimensions, if
provided.}

\item{meta}{an optional named list of whole-plate descriptors}
}
\value{
a \code{Plate} object
}
\description{
Plate constructor
}
