% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plate_formats}
\alias{plate_formats}
\title{Standard plate formats for microwell plates}
\format{
An object of class \code{data.frame} with 9 rows and 3 columns.
}
\usage{
plate_formats
}
\description{
Standard plate formats for microwell plates
}
\keyword{datasets}
