% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.gp}
\alias{print.gp}
\title{Print a gp object}
\usage{
\method{print}{gp}(x, ...)
}
\arguments{
\item{x}{A gp object}

\item{...}{Additional parameters to pass to \code{print}}
}
\description{
Print a gp object
}
\keyword{internal}
