% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein_quant}
\alias{protein_quant}
\title{Data from a protein quantification assay in a 96-well plate. See the "Using
gp to wrangle plate data" vignette for more details}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 8 rows and 12 columns.
}
\usage{
protein_quant
}
\description{
Data from a protein quantification assay in a 96-well plate. See the "Using
gp to wrangle plate data" vignette for more details
}
\keyword{datasets}
