% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{scrape}
\alias{scrape}
\title{Scrape object generic}
\usage{
scrape(x, positions = NULL, negate = FALSE)
}
\arguments{
\item{x}{Object to have layers removed from it}

\item{positions}{numeric or character vector specifying the positions or names of layers to be removed.}

\item{negate}{logical. If TRUE, values will be kept instead of removed.}
}
\value{

}
\description{
Scrape object generic
}
