% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\name{spill}
\alias{spill}
\title{Layer accessor generic}
\usage{
spill(x, n = NULL)
}
\arguments{
\item{x}{Object to be accessed}

\item{n}{optional, character name or numeric position of layer to access}
}
\value{

}
\description{
Layer accessor generic
}
