% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec_prototypes.R
\name{unroll_sec_dim_along_parent}
\alias{unroll_sec_dim_along_parent}
\title{Repeat one dimension of a child section across a dimension of a parent section}
\usage{
unroll_sec_dim_along_parent(gp, dim, wrap)

unroll_sec_dim_along_parent(gp, dim, wrap)
}
\arguments{
\item{gp}{A \code{gp}}

\item{dim}{Either "row" or "col"}
}
\value{
A \code{gp}

A \code{gp}
}
\description{
Repeat one dimension of a child section across a dimension of a parent section

Repeat one dimension of a child section across a dimension of a parent section
}
