% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{well_data}
\alias{well_data}
\alias{well_data.gp}
\title{Extract well data from an object}
\usage{
well_data(x, ...)

\method{well_data}{gp}(x, ...)
}
\arguments{
\item{x}{An object to be passed to its respective method}

\item{...}{Additional arguments (unused)}
}
\value{
A \code{tibble} containing the well data of the object
}
\description{
Extract well data from an object
}
