% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{wells<-}
\alias{wells<-}
\title{Wells setter generic}
\usage{
wells(x) <- value
}
\arguments{
\item{x}{Object to set wells.}

\item{value}{numeric. Number of wells to set.}
}
\value{
x
}
\description{
Wells setter generic
}
