% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wells.R
\name{wells}
\alias{wells}
\title{Wells accessor generic}
\usage{
wells(x)
}
\arguments{
\item{x}{Object to access wells from.}
}
\value{
numeric, number of wells in the object.
}
\description{
Wells accessor generic
}
