% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_plot.R
\name{gp_plot}
\alias{gp_plot}
\title{Plot a \code{gp} object}
\usage{
gp_plot(gp, name = .sec)
}
\arguments{
\item{gp}{A \code{gp} object}

\item{name}{Symbol. Name of a column in \code{gp$well_data} to use as a color.}
}
\value{
a \code{ggplot}
}
\description{
Plot a \code{gp} object
}
\examples{

gp(16, 24) |> gp_plot(.row)
}
