% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_sec.R
\name{gp_sec}
\alias{gp_sec}
\title{Add a section to a \code{gp} object.}
\usage{
gp_sec(
  gp,
  name,
  nrow = NULL,
  ncol = NULL,
  labels = NULL,
  start_corner = c("tl", "tr", "bl", "br"),
  flow = c("row", "col"),
  margin = 0,
  wrap = FALSE,
  break_sections = TRUE,
  advance = TRUE
)
}
\arguments{
\item{gp}{A \code{gp} object}

\item{name}{Character. Name of the section.}

\item{nrow}{Numeric. Section height. If \code{NULL}, will fill width of parent
section.}

\item{ncol}{Numeric. Section width. If \code{NULL}, will fill width of parent
section.}

\item{labels}{Optional. What should the labels of each section be?}

\item{start_corner}{Corner of section to place first item.}

\item{flow}{Direction that subsequent items should be placed relative to first corner.}

\item{margin}{Border width outside the section that will be unfilled. Can
take an argument of one (same border all around), two (top/bottom,
left/right), three (top, left/right, bottom), or four (top, right, bottom,
left).}

\item{wrap}{Should the sections that go off the edge continue on the next row/column?}

\item{break_sections}{Should partial sections be allowed?}

\item{advance}{Should this section be a child or sibling of the one before it? If TRUE (default), it will be a child.}
}
\value{
a \code{gp}
}
\description{
Add a section to a \code{gp} object.
}
\examples{

gp(16, 24) |> gp_sec("section 1", ncol = 3)

pq <- gp(8, 12, protein_quant) |> gp_sec("has_sample", 3, 19, wrap = TRUE, labels = "sample")

# Sections can be used to label things for tidying
pq |> gp_serve()

# They can also be used for plotting:
pq |> gp_plot(has_sample)

}
