% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_sec.R
\name{gp_sec}
\alias{gp_sec}
\title{Add a section to a \code{gp} object.}
\usage{
gp_sec(
  gp,
  name,
  nrow = NULL,
  ncol = NULL,
  start_corner = c("tl", "tr", "bl", "br"),
  flow = c("row", "col"),
  padding = 0,
  margin = 0,
  wrap = FALSE
)
}
\arguments{
\item{gp}{a \code{gp} object}

\item{name}{Character. Name of the section.}

\item{nrow}{Numeric. Section height. If \code{NULL}, will fill width of parent
section.}

\item{ncol}{Numeric. Section width. If \code{NULL}, will fill width of parent
section.}

\item{start_corner}{Corner of section to place first item.}

\item{flow}{Direction that subsequent items should be placed relative to first corner.}

\item{padding}{Border width INSIDE the section that will be unfilled. Can
take an argument of one (same border all around), two (top/bottom,
left/right), three (top, left/right, bottom), or four (top, right, bottom,
left).}

\item{margin}{Border width OUTSIDE the section that will remain unfilled. See
above for possible argument lengths.}

\item{break_sections}{If a section would need to be wrapped around the edge
of the parent section (default parent section is the plate), should it
instead shift the flow direction far enough so that it can fit without
wrapping? Will error if either \code{nrow} or \code{ncol} exceeds the parent
dimension.}
}
\value{

}
\description{
Add a section to a \code{gp} object.
}
