% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_sec.R
\name{gp_sec}
\alias{gp_sec}
\title{Add a section to a \code{gp} object.}
\usage{
gp_sec(
  gp,
  name,
  nrow = NULL,
  ncol = NULL,
  start_corner = c("tl", "tr", "bl", "br"),
  flow = c("row", "col"),
  margin = 0,
  wrap = FALSE,
  break_sections = TRUE
)
}
\arguments{
\item{gp}{a \code{gp} object}

\item{name}{Character. Name of the section.}

\item{nrow}{Numeric. Section height. If \code{NULL}, will fill width of parent
section.}

\item{ncol}{Numeric. Section width. If \code{NULL}, will fill width of parent
section.}

\item{start_corner}{Corner of section to place first item.}

\item{flow}{Direction that subsequent items should be placed relative to first corner.}

\item{margin}{Border width outside the section that will be unfilled. Can
take an argument of one (same border all around), two (top/bottom,
left/right), three (top, left/right, bottom), or four (top, right, bottom,
left).}

\item{break_sections}{Should partial sections be allowed?}
}
\value{

}
\description{
Add a section to a \code{gp} object.
}
