% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_unravel.R
\name{gp_unravel}
\alias{gp_unravel}
\title{Turn data from 'plate form' to 'tidy form'}
\usage{
gp_unravel(df, rownames = NULL)
}
\arguments{
\item{rownames}{Optional character. If there is a colname that specifies the
row index, it will be arranged by this column, then dropped.}
}
\value{
a tibble
}
\description{
Turn data from 'plate form' to 'tidy form'
}
\examples{

test_plate <- matrix(sample(1:10, 96, replace = TRUE), nrow = 8, ncol = 12)

gp_unravel(test_plate)
}
