% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_unravel.R
\name{gp_unravel}
\alias{gp_unravel}
\alias{gp_reravel}
\title{Turn data from 'plate form' to 'tidy form' and back}
\usage{
gp_unravel(df, rownames = NULL)

gp_reravel(df, row_name = ".row", col_name = ".col", values = "value")
}
\arguments{
\item{df}{The \code{data.frame} in plate-form to be tidied}

\item{rownames}{Optional character. If there is a colname that specifies the
row index, it will be arranged by this column, then dropped.}
}
\value{
a tibble
}
\description{
Turn data from 'plate form' to 'tidy form' and back
}
\examples{

test_plate <- matrix(sample(1:10, 96, replace = TRUE), nrow = 8, ncol = 12)

gp_unravel(test_plate)
}
