% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp.R
\name{wells}
\alias{wells}
\alias{wells.gp}
\title{Calculate wells of gp object}
\usage{
wells(x, ...)

\method{wells}{gp}(x, ...)
}
\arguments{
\item{x}{a \code{gp}}

\item{...}{Unused}
}
\description{
Calculate wells of gp object
}
